Fix incorrect service ID
