+++ ScienceJournal/BLEArduino/ScienceKit Sensors/BLEScienceKitAccelerometerXSensor.swift
