+++ ScienceJournal/BLEArduino/ScienceKit Sensors/BLEScienceKitCurrentSensor.swift
