+++ ScienceJournal/BLEArduino/ScienceKit Sensors/BLEScienceKitGyroscopeXSensor.swift
