+++ ScienceJournal/BLEArduino/ScienceKit Sensors/BLEScienceKitGyroscopeYSensor.swift
