+++ ScienceJournal/BLEArduino/ScienceKit Sensors/BLEScienceKitGyroscopeZSensor.swift
