+++ ScienceJournal/BLEArduino/ScienceKit Sensors/BLEScienceKitInput1Sensor.swift
