+++ ScienceJournal/BLEArduino/ScienceKit Sensors/BLEScienceKitInput2Sensor.swift
