+++ ScienceJournal/BLEArduino/ScienceKit Sensors/BLEScienceKitInput3Sensor.swift
