+++ ScienceJournal/BLEArduino/ScienceKit Sensors/BLEScienceKitLinearAccelerometerSensor.swift
