+++ ScienceJournal/BLEArduino/ScienceKit Sensors/BLEScienceKitMagnetometerSensor.swift
