+++ ScienceJournal/BLEArduino/ScienceKit Sensors/BLEScienceKitResistanceSensor.swift
