+++ ScienceJournal/BLEArduino/UI/ScienceKitSensorConfigSelectorView.swift
